Added pathing algorithm
It consists of finding the aisle and row of a certain shelf item. It will adjust based on the location and move accordingly.

Customer first moves up and prioritizes rows. When in a row, find the aisle.

Otherwise, when in an aisle and item not in the aisle, move out into a row and find the aisle.